extract useNavSetup hook
