+++ modules/ui/src/components/Nav/Nav.tsx
@@ -1,11 +1,5 @@
 import classNames from "classnames";
-import React, {
-  FunctionComponent,
-  useMemo,
-  useState,
-  useEffect,
-  Fragment,
-} from "react";
+import React, { FunctionComponent, useMemo, useEffect, Fragment } from "react";
 import { AlgolNav, AppActions } from "../../../../types";
 import css from "./Nav.cssProxy";
 import { NavBottomRow } from "./Nav.BottomRow";
@@ -18,6 +12,7 @@ import { findShortcut } from "../../../../common/nav/findShortcut";
 import { DASHED_SHORTCUTS } from "./Nav.constants";
 import { NavHomeButton } from "./Nav.HomeButton";
 import { NavToggleButton } from "./Nav.ToggleButton";
+import { useNavState } from "./Nav.useNavSetup";
 
 export type NavProps = {
   nav?: AlgolNav;
@@ -29,22 +24,8 @@ const BACK_BUTTON = true;
 const prefetched: Record<string, boolean> = {};
 
 export const Nav: FunctionComponent<NavProps> = props => {
-  const [fullNav, _setFullNav] = useState(false);
-  const [neverNav, _setNeverNav] = useState(true);
-  const setFullNav = (bool: boolean) => {
-    if (bool && neverNav) {
-      _setNeverNav(false);
-    }
-    _setFullNav(bool);
-  };
   const { nav, actions: _actions } = props;
-  const actions = useMemo(
-    () => ({
-      ..._actions,
-      setFullNav,
-    }),
-    [_actions, setFullNav, neverNav]
-  );
+  const { actions, fullNav, neverNav } = useNavState(_actions);
   if (!nav) return <div></div>;
   const { crumbs, me } = nav;
   const hasCrumbs = Boolean(nav && crumbs.length > 0);
@@ -52,7 +33,7 @@ export const Nav: FunctionComponent<NavProps> = props => {
   const shortcut = useMemo(() => findShortcut(nav), [nav]);
 
   useEffect(() => {
-    setFullNav(false);
+    actions.setFullNav(false);
     if (nav) {
       const allSteps = nav.crumbs.concat(nav.me).flatMap(s => [s, ...s.links]);
       if (nav.me.url) {
